+++ llvm/lib/Target/RISCV/RISCVSubtarget.h
@@ -207,9 +207,7 @@ public:
 
   bool hasBEXTILike() const { return HasStdExtZbs || HasVendorXTHeadBs; }
 
-  bool hasCZEROLike() const {
-    return HasStdExtZicond || HasVendorXVentanaCondOps;
-  }
+  bool hasCZEROLike() const { return HasStdExtZicond; }
 
   bool hasConditionalMoveFusion() const {
     // Do we support fusing a branch+mv or branch+c.mv as a conditional move.